chore: Move utils tests to the separate folder
