+++ projects/hslayers/test/add-data-url.spec.ts
@@ -29,7 +29,7 @@ import {HsUrlWfsModule} from 'hslayers-ng/components/add-data';
 import {HsUrlWmsModule} from 'hslayers-ng/components/add-data';
 import {HsUrlWmtsModule} from 'hslayers-ng/components/add-data';
 import {HsUtilsService} from 'hslayers-ng/services/utils';
-import {HsUtilsServiceMock} from './utils.service.mock';
+import {HsUtilsServiceMock} from './utils/utils.service.mock';
 import {HsWmsGetCapabilitiesService} from 'hslayers-ng/services/get-capabilities';
 import {mockLayerUtilsService} from './layer-utils.service.mock';
 import {testingServiceEndpoints} from './data/service-endpoints';
