+++ projects/hslayers/test/map.spec.ts
@@ -24,7 +24,7 @@ import {HsMapService} from 'hslayers-ng/services/map';
 import {HsShareUrlService} from 'hslayers-ng/services/share';
 import {HsShareUrlServiceMock} from './share-url.service.mock';
 import {HsUtilsService} from 'hslayers-ng/services/utils';
-import {HsUtilsServiceMock} from './utils.service.mock';
+import {HsUtilsServiceMock} from './utils/utils.service.mock';
 import {HslayersService} from 'hslayers-ng/core';
 
 class emptyMock {
